Consolidating middleware
