+++ sdk/highlight-next/src/server.edge.ts
@@ -1,10 +1,10 @@
 export type { HighlightEnv } from './util/types'
 
 import * as withHighlightEdge from './util/with-highlight-edge'
-import type { NextRequest } from 'next/server'
 
 import type { HighlightEnv } from './util/types'
 export { H } from './util/highlight-edge'
+export { highlightMiddleware } from './util/highlight-middleware'
 
 export function registerHighlight() {}
 
@@ -24,12 +24,4 @@ export function PageRouterHighlight(_: HighlightEnv) {
 
 export function AppRouterHighlight(_: HighlightEnv) {
 	throw new Error('Do not use AppRouterHighlight() in the edge runtime.')
-}
-
-export function highlightMiddleware(request: NextRequest) {
-	const sessionSecureID = request.cookies.get('sessionSecureID')?.value
-	const xHighlightRequest = request.headers.get('x-highlight-request')
-	if (!xHighlightRequest && sessionSecureID) {
-		request.headers.set('x-highlight-request', sessionSecureID)
-	}
 }
