+++ sdk/highlight-next/src/util/highlight-middleware.ts
@@ -1,0 +1,9 @@
+import type { NextRequest } from 'next/server'
+
+export function highlightMiddleware(request: NextRequest) {
+	const sessionSecureID = request.cookies.get('sessionSecureID')?.value
+	const xHighlightRequest = request.headers.get('x-highlight-request')
+	if (!xHighlightRequest && sessionSecureID) {
+		request.headers.set('x-highlight-request', sessionSecureID)
+	}
+}
